sprint-4/step-1 - ProfileEditor extract
