+++ src/components/profile-editor/profile-editor.module.css
@@ -1,0 +1,17 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    gap: 24px;
+}
+
+
+.buttons {
+    display: flex;
+    align-items: center;
+    gap: 28px;
+}
+
+.hidden {
+    visibility: hidden;
+}
