+++ src/pages/profile/profile.module.css
@@ -43,12 +43,3 @@
     color: #4C4CFF;
 }
 
-.buttons {
-    display: flex;
-    align-items: center;
-    gap: 28px;
-}
-
-.hidden {
-    visibility: hidden;
-}
